Refactor everything into async/await
This patch refactors the complete codebase to not use

    `tokio::runtime::Handle::current().block_on()`

but use async/await syntax everywhere.

One small thing, though, had to be changed to actually use the
block_on() call:

     ItemsIter::fill_next_page()

added a blocking call, because refactoring that into a proper Stream
would make the size of this commit explode. To keep the review managable
for this commit, we introduced the blocking call here, so a huge
improvement with async/await instead of blocking everywhere and this
one blocking call introduced instead.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>